[refactor] 機能をhookに分割
